+++ nengo/__init__.py
@@ -1,3 +1,5 @@
+# pylint: disable=wrong-import-order,wrong-import-position
+
 """
 Nengo
 =====
@@ -10,8 +12,29 @@ in the `examples` directory of the source code repository.
 """
 
 import logging
+import sys
 
 from .version import version as __version__
+
+if sys.version_info < (3, 5):
+    raise ImportError("""
+You are running Python version %s with Nengo version %s.
+Nengo requires at least Python 3.5.
+
+The fact that this version was installed on your system probably means that you
+are using an older version of pip; you should consider upgrading with
+
+ $ pip install pip setuptools --upgrade
+
+There are two options for getting Nengo working:
+
+- Upgrade to Python >= 3.5
+
+- Install an older version of Nengo:
+
+ $ pip install 'nengo<3.0'
+""" % (sys.version, __version__))
+del sys
 
 # Nengo namespace (API)
 from .base import Process